To improve the text, consider the following revised version:
"We have now implemented a signup and login page. Initially, we created a forms.py file in the base directory and added both signup and login forms to it. The login form was customized differently from the signup form.

For the signup functionality, we added a view with functional code. However, for the login process, we incorporated an authentication module. This involved using a generic view and passing the created forms as arguments, specifying the template name accordingly.

Additionally, modifications were made to the base.html file to accommodate these changes."